src/main/python/allocate_ip/requirements.txt
requests==2.21.0
dnspython
